Mads review fixes
