+++ source/OutgoingMessages.Domain/DocumentWriters/DocumentFactory.cs
@@ -17,12 +17,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-using Energinet.DataHub.EDI.OutgoingMessages.Domain.DocumentWriters;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.MarketDocuments;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.OutgoingMessages;
 using NodaTime;
 
-namespace Energinet.DataHub.EDI.OutgoingMessages.Domain;
+namespace Energinet.DataHub.EDI.OutgoingMessages.Domain.DocumentWriters;
 
 public class DocumentFactory
 {
