+++ source/OutgoingMessages.Domain/MarketDocuments/MarketDocument.cs
@@ -17,7 +17,7 @@ using System.Diagnostics.CodeAnalysis;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.Models;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.ActorMessagesQueues;
-using Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.Bundels;
+using Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.Bundles;
 
 namespace Energinet.DataHub.EDI.OutgoingMessages.Domain.MarketDocuments;
