+++ source/OutgoingMessages.Domain/Models/Bundles/Bundle.cs
@@ -16,7 +16,7 @@ using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 using Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.OutgoingMessages;
 using NodaTime;
 
-namespace Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.Bundels;
+namespace Energinet.DataHub.EDI.OutgoingMessages.Domain.Models.Bundles;
 
 public sealed class Bundle
 {
